WIP dataframe integratinon
