+++ ts/packages/memory/image/src/index.ts
@@ -3,3 +3,6 @@
 
 export * from "./importImages.js";
 export * from "./indexingService.js";
+export * from "./imageCollection.js";
+export * from "./tables.js";
+export * from "./imageMeta.js";
