+++ ts/pnpm-workspace.yaml
@@ -6,11 +6,13 @@ packages:
   - examples/*
   - examples/agentExamples/*
   - tools
+
 onlyBuiltDependencies:
-  - "@azure/msal-node-extensions"
-  - "@azure/msal-node-runtime"
-  - "@vscode/vsce-sign"
+  - '@azure/msal-node-extensions'
+  - '@azure/msal-node-runtime'
+  - '@vscode/vsce-sign'
   - better-sqlite3
+  - canvas
   - electron
   - esbuild
   - exifreader
